DDF-2165 Addressing PR Comments
Splitting the Eventhandler from the action provider
